requirements.txt
requests
django
django-rest-framework
gunicorn
psycopg2
